fix: 🐛 ensure type exactly converted

+++ src/main/java/uk/ac/warwick/dcs/sherlock/module/model/base/scoring/NGramScorer.java
@@ -132,7 +132,7 @@ public class NGramScorer {
      */
     public boolean checkSize(int file_count, ArrayList<NgramMatch> list) {
         // if the match is uncommon return true
-        if ((file_list.size() / file_count) <= threshold) {
+        if (((float) file_list.size() / file_count) <= threshold) {
             return true;
         }
         // if the match is common set all pairs in the list to be common
